Fix maven-enforcer usage
Old logic did get rid of warning by causing an error.  This corrects
that issue.  Sonatype overwrites the version due to using it in
unmanaged location.  This is ultimately probably a maven bug.  So maven
will warn us that we are overriding same version but we have to in order
to properly enforce version 1.4 of enforce plugin.  This isn't true on
all uses of this so it's even more probable it is a maven bug.