Turn off validate restriction on build
Having validate enforced requires that users format some other way
specifically using eclipse which violates how this works in general.  We
format what we are working on with the plugin not force users to use
other tools to do that.  It defeats the process.  Our CI process should
perform validate and lock it if things are not validated.