Remove unnecessary SystemUtil class
The SystemUtil class previously contained a String constant for the
system line ending. This isn't needed as of Java 7, which provides this
method for us as System.lineSeparator()

In the git history, there is a mention of this class being introduced to
fix a problem. However, I cannot determine what problem it could have
fixed, and I cannot find any problem using the new Java 7 method.