[revert-test] Remove enabled if system property - it already checks that internally
This test is intended to run on all platforms.  Unless there is some oddity on a platform that uses different line endings than *nix / windows (ie is null for some reason), this would work.  This seems like someone was having issues on their machine likely to do some mocking out of system line endings if I had to guess.