[enhance] Do not write the cache properties file unless it actually changed
For those that want to save the cache as it does survive due to pathing differences on both windows/*nix, its impossible to properly save as java properties will apply a timestamp into the file.  To avoid this, simply don't write the file unless we needed to as it was wasteful write anyways.