[javadoc] Add javadoc indicating usage of cache
cache intent at the moment is it live past the clean cycle of maven.  There is no benefit to running formatting when unnecessary even if just cleaning project.  This brings speed of plugin down considerably by skipping the formatter all together even on first build.  Users can include the cache to git or they can ignore it or they can change it.

There are current 2 requests to keep as it was originally in target.  This is being monitored to see if there is a going concenses as to default behaviour.  The original problem was that the cache was not deterministic, files were read too many times, written too many times, and it was unclear about the caching.  This makes it absolutely clear.  If enough people perfer old hiding of it in target, that will be accommodated.  I've tested it on large scale projects and teams using it haven't had any concerns with the new way so its too early to tell if this is going to stick or change.