Change log level for "<type> formatting is skipped" messages
If you skip certain types of files from formatting and your project
have some files of those types then one message per file saying "<type>
formatting is skipped" is written on the maven console log providing no
value (in part because it does not specify which file was skipped as it
is shown on debug level).

This change put those logs at the same log level (debug) as the one that
shows which file is being processed so you have the same information but
with the proper context and only when you want to debug some case.