[ci] Add some clarifications to jsoup patch processing
This is to address question on code review around line endings.  Added clarity throughout to help assist to why things exist and that normalization of line endings happens early so each call that does not do that now points back to where it was done.

Effectively we take formattedCode, split it, do something with it, then recreate formattedCode so the \\r\\n only needs to occur once as its not possible afterwards.  The issue originates in jsoup and only needs the one pass.

Further out builds run two times across these, if the first was to be changed, it immediately fails.  So its been vetted and checked but certainly wording was not entirely clear.