Update fluido skin and require JDK 17 to build
Building the site does not work in 2.22.0 without these changes, because
of the updates to the tycho transitive dependencies that are analyzed
during the build of the site, and because of constraints on the version
of doxia for the site build.

This change bumps fluido to the latest, and makes JDK 17 the minimum
version required to build, so these restrictions are explicit in future
releases.