[caching] Rewrite d1d5852 fixing #757
Original change in d1d5852 only addressed java and did so at detriment of all formatters.  This caused certain platforms to continually update the cache file on every build.  As seen on GitHub, Linux and Macos would change the *nix hashes for every file constantly including java.  For Windows on Github, it was changing the entire file (not clear as to why entire file but that was observed over and over).  Testing to show issue was done with 2.23.0 formatter release that introduced the problem.