[var] since majority of 'var' is used, used the rest
'var' is not hard to understand.  Its already stating directly what it is by the right hand side.  This is still type safe.  Further its just the opposite of diamond operation.  Java provided this for a valid reason, we end up with less imports, cleaner code, and lose nothing using it.