[pom] Site still did not release, add scm data to allow it to occur
actual issue is all the overrides on release that are not needed.  Fall to scm as that is the standard way maven suggests now even though release wants site.  The release still won't do it but left instructiosn on here to do this on demand.  Takes maybe minute or two.